Build universal binaries for OpenSSL.
Unfortunately, simply passing `CFLAGS=-arch arm64 -arch x86_64` to the
OpenSSL configuration fails.  My guess is there that there are if-defs
keyed off of the toolchain that the `config` script guesses that muck up
the double-arch compile.  So, rework the OpenSSL build steps to produce
universal binaries by building OpenSSL twice, once for arm64 and once
for x86_64, then using `lipo` to stitch the two sets of libraries
together.

Duplicate the previous set of OpenSSL targets, moving the output of one
set to `$(TMP)/openssl/arm64` and the other to `$(TMP)/openssl/x86_64`.

Change the OpenSSL configuration steps to call `openssl/Configure` with
the appropriate toolchain name, instead of the `openssl/config` script,
which guesses the toolchain.

Add targets for the fat binaries made by `lipo` and to copy over the
header files from the arm64 OpenSSL version.