Fix `wget` error "Unable to locally verify the issuer's authority."
Add `--openssldir=` to the OpenSSL configuration options and set it to
point to `/etc/ssl` so that `wget` can find the Apple-provided list of
certificate authorities.  Also remove the `no-stdio` option from the
OpenSSL configuration options, since this prevents OpenSSL from reading
the certificate authorities list.

Added a test to the `check` target to verify that `wget` can process an
HTTPS URL without error.