Add `notarize` target to the makefile.
Add the steps to notarize the signed installer package.  The DAG for
these targets looks like:

        wget-$(version).pkg
            --> $(TMP)/submit-log.json
            --> $(TMP)/submission-id.txt
            --> $(TMP)/notarization-log.json
            --> notarized.stamp.txt
            --> stapled.stamp.txt

These steps track submission to the notary service, extracting the
submission ID, getting a log of notarization details, checking if the
package was accepted and stapling the notarization ticket to the
package.

Also updated the readme file with a section on build prerequisites, info
on the `NOTARIZATION_KEYCHAIN_PROFILE` makefile variable, info on the
notarization target and a section about signing and notarization
credentials.