Bump revision number to 2.
Set the installer package revision to 2.  Update the macOS and Xcode
versions in the readme.  Update how the makefile builds the version,
adding `$(rev)` and `$(ver)` variables; the `$(ver)` variable only
contains the `-rX` suffix if the revision is greater than one.  Update
the final package filename to use `$(ver)`.