Statically link zlib to the `wget` executable.
Added the zlib-1.2.11 distribution to the project.  Added build steps to
build zlib for arm64 and x86_64.

Added `ZLIB_CFLAGS` and `ZLIB_LIBS` to the configure options for wget to
link to the project's static `libz` instead of macOS's dynamic `libz`.

Updated `make check` to check the architectures built for `libiconv` and
`libz`.  Fixed the configure options for `libiconv` to include the
`-arch` flags for arm64 and x86_64.