Combine the wget install and codesigning targets.
The codesign tool modifies the binary in place, which causes problems
for correct calculation of out-of-date targets.  Fix this by merging the
codesigning action into the wget install target so that the `wget`
executable is signed immediately after being installed in the `install`
directory.

Also tweak the related pkg targets to make a cleaner dependency tree,
with `wget.pkg` depending on `uninstall-wget`, which depends on `wget`
and `wget.path` being installed, so that the tree of installed files can
be correctly built.