Fix dependency resolution of the `notarize` target.
The `stapler` command modifies the package in place, which breaks
out-of-date calculation.  Modify the chain of targets leading to
`notarize` to fix this by building the unnotarized package in the temp
directory as `wget-$(ver)-unnotarized.pkg` and replacing the
`stapled.stamp.txt` target with the `wget-$(ver).pkg` target in the
project root, changing the action to first copy the unnotarized package
then staple it.